requirements.txt
python-dotenv==1.0.1
aiogram==3.3.0
youtube-transcript-api==0.6.1
requests==2.31.0
openai==1.12.0
aiohttp==3.9.3
loguru==0.7.2 